:recycle: Separate utility files
